api/requirements.txt
django==3.0.8
djangorestframework==3.11
djangorestframework-simplejwt==4.4
django-cors-headers==3.4
psycopg2-binary==2.8
python-dotenv==0.14
celery==4.4
django-reversion==3.0.7
pillow==7.2.0


flake8==3.8.3
mypy
pyflakes==2.2.0
pylint==2.6.0
pycodestyle==2.6.0
ipython==7.17.0
ptvsd==4.3.2
django_extensions
